an action may result in subscription of type  SubscriptionType
Some actions registered on instrumented code  could have as side effect that tool clients will periodically be provided some valuel.
For instance, in the case of an inspect state action, the client will be periodically provided with the inspected state.
Those actions are ones for which the SubscriptionType and methods `parseSubscriptionData` and `onSubscriptionData` are required;